agent/voice-ai/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lead360-voice-ai-agent"
version = "1.0.0"
description = "Lead360 Voice AI Agent using LiveKit Agents SDK"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=0.8.0",
    "livekit-plugins-deepgram>=0.6.0",
    "livekit-plugins-openai>=0.8.0",
    "livekit-plugins-cartesia>=0.4.0",
    "livekit-plugins-silero>=1.4.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",  # httpx mock library
]

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
